transferring lib ownership to k8s charm (#11)
# PR 
## Proposal
The pgbouncer charm lib previously lived on the VM charm. Since I'm planning to use the k8s charm for development moving forward, porting implementation over to the VM charm once it's fully implemented, tested, and reviewed, it makes sense to keep the charm library on the k8s charm as well. 

## Context
- I'm planning to use the k8s charm for development because it builds and deploys faster, making testing and development faster. 
- All this code has already been reviewed on the VM charm, I'm just changing the names of things and copying unit tests over. 

## Release Notes
- Rename pgb lib and copy unit tests over from VM charm.

## Testing
- Unit and integration tests have run

## Commits
* transferring lib ownership to k8s charm

* renamed application

* updated app name

* added dev test marker