Fix deletion test (#275)
Removes the need for patch_port() at pebble-ready event. There is no reason to check if we need to patch the port at this event, as port should be toggled to ClusterIP / NodePort in response to client relations exclusively. Also, pebble-ready happens at every restart of the pod, which per-se does not mark as a change on exposing or not the service.

Simplifying the external_connectivity() to make it more readable and adding extra safeguards in the relation checking methods for node-port.

Fixes the tests to consider NodePort or ClusterIP accordingly.